Replaced tabs with spaces.
GameWindow no longer raises a Resize event before Load. This was a hack that was introduced to work around some design limitations in the early days of the library. These limitations no longer hold (INativeWindow size is now set independently of the Resize event), so the hack is no longer necessary. The To avoid breaking existing applications, a Resize event is still raised unconditionally after Load is complete.